Move RequestTracker into the client-api package
