+++ servicetalk-grpc-netty/build.gradle
@@ -36,7 +36,6 @@ dependencies {
   implementation project(":servicetalk-transport-netty-internal")
   implementation project(":servicetalk-utils-internal")
   implementation project(":servicetalk-concurrent-internal")
-  implementation project(":servicetalk-loadbalancer")
   implementation "org.slf4j:slf4j-api:$slf4jVersion"
   implementation "com.google.protobuf:protobuf-java"
