+++ servicetalk-http-netty/src/test/java/io/servicetalk/http/netty/HttpRequestTrackerTest.java
@@ -16,6 +16,7 @@
 package io.servicetalk.http.netty;
 
 import io.servicetalk.client.api.DelegatingConnectionFactory;
+import io.servicetalk.client.api.RequestTracker;
 import io.servicetalk.concurrent.api.Single;
 import io.servicetalk.context.api.ContextMap;
 import io.servicetalk.http.api.FilterableStreamingHttpConnection;
@@ -23,7 +24,6 @@ import io.servicetalk.http.api.HttpClient;
 import io.servicetalk.http.api.HttpResponse;
 import io.servicetalk.http.api.HttpService;
 import io.servicetalk.http.api.SingleAddressHttpClientBuilder;
-import io.servicetalk.loadbalancer.RequestTracker;
 import io.servicetalk.transport.api.HostAndPort;
 import io.servicetalk.transport.api.ServerContext;
 import io.servicetalk.transport.api.TransportObserver;
@@ -39,7 +39,7 @@ import javax.annotation.Nullable;
 
 import static io.netty.handler.codec.http.HttpStatusClass.SERVER_ERROR;
 import static io.netty.handler.codec.http.HttpStatusClass.SUCCESS;
-import static io.servicetalk.loadbalancer.RequestTracker.REQUEST_TRACKER_KEY;
+import static io.servicetalk.client.api.RequestTracker.REQUEST_TRACKER_KEY;
 import static io.servicetalk.transport.netty.internal.AddressUtils.localAddress;
 import static io.servicetalk.transport.netty.internal.AddressUtils.serverHostAndPort;
 import static org.hamcrest.MatcherAssert.assertThat;
