+++ servicetalk-loadbalancer/src/test/java/io/servicetalk/loadbalancer/XdsOutlierDetectorAlgorithmTest.java
@@ -15,6 +15,7 @@
  */
 package io.servicetalk.loadbalancer;
 
+import io.servicetalk.client.api.RequestTracker;
 import io.servicetalk.concurrent.api.ExecutorExtension;
 import io.servicetalk.concurrent.api.TestExecutor;
