+++ servicetalk-loadbalancer/src/test/java/io/servicetalk/loadbalancer/XdsOutlierDetectorTest.java
@@ -15,6 +15,7 @@
  */
 package io.servicetalk.loadbalancer;
 
+import io.servicetalk.client.api.RequestTracker;
 import io.servicetalk.concurrent.api.TestExecutor;
 
 import org.junit.jupiter.api.Test;
